// vendor/node-worker/src/worker/epoxy/epoxy-wasm.d.ts
declare class WsReadEvent {}
declare class WsWriteEvent {}
// ...
type EpoxyRawHeaders = Record<string, string[]>;
type EpoxyResponse = Response & {
    rawHeaders: EpoxyRawHeaders;
};

type EpoxyWSChunk = string | Uint8Array<ArrayBufferLike>;
interface EpoxyWSCloseInfo {
    closeCode?: number;
    reason?: string;
}
interface EpoxyWebSocketOptions {
    protocols?: string | string[];
    headers?: HeadersInit;
}
declare class EpoxyWS {
    #private;
    readonly readable: ReadableStream<EpoxyWSChunk>;
    readonly writable: WritableStream<EpoxyWSChunk>;
    readonly protocol: string;
    readonly headers: Headers;
    readonly rawHeaders: EpoxyRawHeaders;
    readonly closed: Promise<EpoxyWSCloseInfo>;
    constructor(readable: ReadableStream<WsReadEvent>, writable: WritableStream<WsWriteEvent>, protocol: string, headers: Headers, rawHeaders: EpoxyRawHeaders);
    close(closeInfo?: EpoxyWSCloseInfo): void;
}